One class/one file + highscores
